Generalize NULL-ordering test to a shared cross-backend suite
Lift the MSSQL-specific NullOrderingSuite into modules/sql-core as
SqlNullOrderingMapping/SqlNullOrderingSuite, following the same
pattern as SqlQualifiedNamesMapping/Suite. The expected surviving ids
depend only on the requested nullsLast flag, never on a backend's
native NULL-sort default, so the exact same suite generalizes
unchanged to nulls-high backends (Postgres, Oracle) as well as the
nulls-low ones.

Wires the shared suite into doobie-pg, skunk, doobie-oracle, and
doobie-mssql (replacing its standalone duplicate). sqlite/mysql/
mariadb/h2 aren't reachable from this branch and will adopt the same
shared suite once their branches rebase past this merge.
